+++ R/helpers_model.R
@@ -38,10 +38,7 @@ h_model_dual_endpoint_sigma2W <- function(use_fixed,
         precW ~ dgamma(precWa, precWb)
       }
     )
-    comp$modelspecs <- c(
-      comp$modelspecs,
-      list(precWa = sigma2W[["a"]], precWb = sigma2W[["b"]])
-    )
+    comp$modelspecs <- c(comp$modelspecs, list(precWa = sigma2W[["a"]], precWb = sigma2W[["b"]]))
     comp$init$precW <- 1
     comp$sample <- c(comp$sample, "precW")
   }
@@ -92,10 +89,7 @@ h_model_dual_endpoint_rho <- function(use_fixed,
         rho <- 2 * kappa - 1
       }
     )
-    comp$modelspecs <- c(
-      comp$modelspecs,
-      list(rhoa = rho[["a"]], rhob = rho[["b"]])
-    )
+    comp$modelspecs <- c(comp$modelspecs, list(rhoa = rho[["a"]], rhob = rho[["b"]]))
     comp$init$kappa <- 0.5
     comp$sample <- c(comp$sample, "rho")
   }
@@ -123,6 +117,7 @@ h_model_dual_endpoint_rho <- function(use_fixed,
 #' @return A [`DualEndpoint`] model with updated `priormodel`, `modelspecs`,
 #'   `init`, `sample` slots.
 #'
+#' @seealso [`DualEndpointRW`].
 #' @export
 h_model_dual_endpoint_sigma2betaW <- function(use_fixed,
                                               sigma2betaW,
@@ -132,8 +127,8 @@ h_model_dual_endpoint_sigma2betaW <- function(use_fixed,
 
   if (use_fixed) {
     assert_number(sigma2betaW, lower = 0 + .Machine$double.xmin, finite = TRUE)
-    de@modelspecs <- function() {
-      c(modelspecs(), list(precBetaW = 1 / sigma2betaW))
+    de@modelspecs <- function(from_prior) {
+      c(modelspecs(from_prior), list(precBetaW = 1 / sigma2betaW))
     }
   } else {
     assert_true(h_test_named_numeric(sigma2betaW, permutation.of = c("a", "b")))
@@ -144,9 +139,9 @@ h_model_dual_endpoint_sigma2betaW <- function(use_fixed,
         precBetaW ~ dgamma(precBetaWa, precBetaWb)
       }
     )
-    de@modelspecs <- function() {
+    de@modelspecs <- function(from_prior) {
       c(
-        modelspecs(),
+        modelspecs(from_prior),
         list(precBetaWa = sigma2betaW[["a"]], precBetaWb = sigma2betaW[["b"]])
       )
     }
@@ -204,33 +199,27 @@ h_model_dual_endpoint_beta <- function(param,
   assert_class(de, "DualEndpoint")
 
   use_fixed <- setNames(is.scalar(param), param_name)
-  ms <- de@modelspecs
+  modelspecs <- de@modelspecs
+  init <- de@init
 
   if (use_fixed) {
-    de@modelspecs <- function() {
-      c(ms(), setNames(list(param), param_name))
+    de@modelspecs <- function(from_prior) {
+      c(modelspecs(from_prior), setNames(list(param), param_name))
     }
   } else {
     assert_character(param_suffix, len = 2, unique = TRUE, any.missing = FALSE)
     assert_function(priormodel)
     param_name2 <- paste0(param_name, param_suffix)
 
-    init <- de@init
     de@priormodel <- h_jags_join_models(
       de@priormodel,
       priormodel
     )
-    de@modelspecs <- function() {
-      c(
-        ms(),
-        setNames(list(param[1], param[2]), param_name2)
-      )
+    de@modelspecs <- function(from_prior) {
+      c(modelspecs(from_prior), setNames(list(param[1], param[2]), param_name2))
     }
     de@init <- function(y) {
-      c(
-        init(y),
-        setNames(list(mean(param)), param_name)
-      )
+      c(init(y), setNames(list(mean(param)), param_name))
     }
     de@sample <- c(de@sample, param_name)
   }